refactor: Split up config loading file into multiple
